@@ -1,37 +1,33 @@
+"use client";
+
 import Link from "next/link";
-import { useRouter } from "next/router";
-import { i18n, useTranslation } from "next-i18next";
-import { useCallback, useEffect } from "react";
+import { useTranslations, useLocale } from "next-intl";
+import { useMemo } from "react";
+import { usePathname, useRouter } from "next/navigation";
 
 export const Header = () => {
+  const t = useTranslations("common");
   const router = useRouter();
-  const { t } = useTranslation("common");
-  const { locale, locales, pathname, asPath, query } = router;
+  const pathname = usePathname();
+  const currentLocale = useLocale();
 
-  // 🔁 Ensure i18n stays in sync with router locale
-  useEffect(() => {
-    if (i18n?.language !== locale) {
-      i18n?.changeLanguage(locale);
-    }
-  }, [locale]);
+  const locales = ["en", "nl"];
 
-  const changeLanguage = useCallback(
-    async (newLocale: string) => {
-      await router.push({ pathname, query }, asPath, {
-        locale: newLocale,
-        scroll: false,
-      });
-      await i18n?.reloadResources?.(newLocale);
-    },
-    [pathname, query, asPath, router]
+  const navItems = useMemo(
+    () => [
+      { href: "/", label: t("nav.home") },
+      { href: "/projects", label: t("nav.projects") },
+      { href: "/contact", label: t("nav.contact") },
+    ],
+    [t]
   );
 
-  const navItems = [
-    { href: "/", label: t("nav.home") },
-    { href: "/projects", label: t("nav.projects") },
-    { href: "/contact", label: t("nav.contact") },
-  ];
-
+  const changeLanguage = (newLocale: string) => {
+    const segments = pathname.split("/");
+    segments[1] = newLocale;
+    const newPath = segments.join("/");
+    router.replace(newPath);
+  };
   return (
     <header className="sticky top-0 z-50 w-full bg-white/90 backdrop-blur border-b border-neutral-200 shadow-sm">
       <nav className="max-w-7xl mx-auto flex items-center justify-between px-6 py-4">
@@ -41,11 +37,12 @@ export const Header = () => {
 
         <ul className="flex gap-2 sm:gap-4 items-center text-sm font-medium text-gray-700">
           {navItems.map((item) => {
-            const isActive = pathname === item.href;
+            const isActive = pathname === `/${currentLocale}${item.href}`;
             return (
               <li key={item.href}>
                 <Link
                   href={item.href}
+                  locale={currentLocale}
                   className={`px-3 py-2 rounded transition ${
                     isActive
                       ? "bg-blue-50 text-blue-700 font-semibold"
@@ -60,11 +57,11 @@ export const Header = () => {
 
           <li>
             <select
-              value={locale}
+              value={currentLocale}
               onChange={(e) => changeLanguage(e.target.value)}
               className="ml-4 cursor-pointer"
             >
-              {locales?.map((locale) => (
+              {locales.map((locale) => (
                 <option key={locale} value={locale}>
                   {locale.toUpperCase()}
                 </option>
