@@ -1,9 +1,13 @@
-import type { NextConfig } from "next";
-import i18nConfig from "./next-i18next.config.js";
+import createNextIntlPlugin from 'next-intl/plugin';
 
-const nextConfig: NextConfig = {
+const withNextIntl = createNextIntlPlugin('./next-intl.config.ts');
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
   reactStrictMode: true,
-  i18n: i18nConfig.i18n, // Only pass the "i18n" subkey
+  experimental: {
+    serverActions: {}, // correct type
+  },
 };
 
-export default nextConfig;
+export default withNextIntl(nextConfig);
