@@ -1,100 +1,126 @@
-export enum WorkMode {
-  Remote = "workMode.remote",
-  OnSite = "workMode.onSite",
-  Hybrid = "workMode.hybrid",
-}
+// Work mode
+export const WorkMode = {
+  Remote: "workMode.remote",
+  OnSite: "workMode.onSite",
+  Hybrid: "workMode.hybrid",
+} as const;
 
-export enum Language {
-  Dutch = "language.dutch",
-  English = "language.english",
-}
+export type WorkMode = (typeof WorkMode)[keyof typeof WorkMode];
 
-export enum Tech {
-  HTML5 = "tech.html5",
-  CSS3 = "tech.css3",
-  JavaScript = "tech.javascript",
-  TypeScript = "tech.typescript",
-  Angular = "tech.angular",
-  React = "tech.react",
-  Vue = "tech.vue",
-  WebComponents = "tech.webComponents",
-  Stencil = "tech.stencil",
-  WCAG = "tech.wcag",
-  Playwright = "tech.playwright",
-  Cypress = "tech.cypress",
-  REST = "tech.rest",
-  OpenAPI = "tech.openapi",
-  JWT = "tech.jwt",
-  Java8 = "tech.java8",
-  Maven = "tech.maven",
-  Gitlabs = "tech.gitlabs",
-  Git = "tech.git",
-  NPM = "tech.npm",
-  Jenkins = "tech.jenkins",
-  Tekton = "tech.tekton",
-  ArgoCD = "tech.argocd",
-  SPLUNK = "tech.splunk",
-  Docker = "tech.docker",
-  SonarQube = "tech.sonarqube",
-  ESLint = "tech.eslint",
-  Storybook = "tech.storybook",
-  Confluence = "tech.confluence",
-  Jira = "tech.jira",
-  MySQL = "tech.mysql",
-  CSharp = "tech.csharp",
-  Neo4J = "tech.Neo4J",
-  Redux = "tech.Redux",
-  WPF = "tech.WPF",
-  MVVM = "tech.MVVM",
-  WebForms = "tech.WebForms",
-  API = "tech.API",
-  Azure = "tech.Azure",
-  MSSQL = "tech.MSSQL",
-  TailwindCSS = "tech.TailwindCSS",
-  NextJS = "tech.NextJS",
-  RxJS = "tech.RxJS",
-  BEM = "tech.BEM",
-  SCSS = "tech.SCSS",
-  SVG = "tech.SVG",
-  AWSLambda = "tech.AWSLambda",
-  GitLabCI = "tech.GitLabCI",
-  Prettier = "tech.Prettier",
-  HeadlessCMS = "tech.HeadlessCMS",
-  ReactQuery = "tech.ReactQuery",
-  Zustand = "tech.Zustand",
-  Jest = "tech.Jest",
-  GitHub = "tech.GitHub",
-  ComponentDesign = "tech.ComponentDesign",
-  Gitlab = "tech.Gitlab",
-  Bitbucket = "tech.Bitbucket",
-  DotNetCore = "tech.DotNetCore",
-  AzureAppServices = "tech.AzureAppServices",
-  AzureCloud = "tech.AzureCloud",
-  EntityFramework = "tech.EntityFramework",
-  Microservices = "tech.Microservices",
-  RestAPI = "tech.RestAPI",
-  ConfigurationDrivenUI = "tech.ConfigurationDrivenUI",
-  DataDrivenUI = "tech.DataDrivenUI",
-  Kubernetes = "tech.Kubernetes",
-  xUnit = "tech.xUnit",
-  AzureDevOps = "tech.AzureDevOps",
-  AngularMaterial = "tech.AngularMaterial",
-  AzureCosmosDB = "tech.AzureCosmosDB",
-  SQLServer = "tech.SQLServer",
-  ReactRouter = "tech.ReactRouter",
-  Scrum = "tech.Scrum",
-  SourceTree = "tech.SourceTree",
-  NHibernate = "tech.NHibernate",
-  MediatR = "tech.MediatR",
-  AspNetMVC = "AspNetMVC",
-  CQRS = "tech.CQRS",
-  MediatorPattern = "tech.MediatorPattern",
-  TSLint = "tech.TSLint",
-  OctopusDeploy = "tech.OctopusDeploy",
-  UIKit = "tech.UIKit",
-  IceLib = "tech.IceLib",
-  IntegrationTesting = "tech.IntegrationTesting",
-}
+// Languages
+export const Language = {
+  Dutch: "language.dutch",
+  English: "language.english",
+} as const;
+
+export type Language = (typeof Language)[keyof typeof Language];
+
+// Tech stack
+export const Tech = {
+  HTML5: "tech.html5",
+  CSS3: "tech.css3",
+  JavaScript: "tech.javascript",
+  TypeScript: "tech.typescript",
+  Angular: "tech.angular",
+  React: "tech.react",
+  Vue: "tech.vue",
+  WebComponents: "tech.webComponents",
+  Stencil: "tech.stencil",
+  WCAG: "tech.wcag",
+  Playwright: "tech.playwright",
+  Cypress: "tech.cypress",
+  REST: "tech.rest",
+  OpenAPI: "tech.openapi",
+  JWT: "tech.jwt",
+  Java8: "tech.java8",
+  Maven: "tech.maven",
+  Gitlabs: "tech.gitlabs",
+  Git: "tech.git",
+  NPM: "tech.npm",
+  Jenkins: "tech.jenkins",
+  Tekton: "tech.tekton",
+  ArgoCD: "tech.argocd",
+  SPLUNK: "tech.splunk",
+  Docker: "tech.docker",
+  SonarQube: "tech.sonarqube",
+  ESLint: "tech.eslint",
+  Storybook: "tech.storybook",
+  Confluence: "tech.confluence",
+  Jira: "tech.jira",
+  MySQL: "tech.mysql",
+  CSharp: "tech.csharp",
+  Neo4J: "tech.Neo4J",
+  Redux: "tech.Redux",
+  WPF: "tech.WPF",
+  MVVM: "tech.MVVM",
+  WebForms: "tech.WebForms",
+  API: "tech.API",
+  Azure: "tech.Azure",
+  MSSQL: "tech.MSSQL",
+  TailwindCSS: "tech.TailwindCSS",
+  NextJS: "tech.NextJS",
+  RxJS: "tech.RxJS",
+  BEM: "tech.BEM",
+  SCSS: "tech.SCSS",
+  SVG: "tech.SVG",
+  AWSLambda: "tech.AWSLambda",
+  GitLabCI: "tech.GitLabCI",
+  Prettier: "tech.Prettier",
+  HeadlessCMS: "tech.HeadlessCMS",
+  ReactQuery: "tech.ReactQuery",
+  Zustand: "tech.Zustand",
+  Jest: "tech.Jest",
+  GitHub: "tech.GitHub",
+  ComponentDesign: "tech.ComponentDesign",
+  Gitlab: "tech.Gitlab",
+  Bitbucket: "tech.Bitbucket",
+  DotNetCore: "tech.DotNetCore",
+  AzureAppServices: "tech.AzureAppServices",
+  AzureCloud: "tech.AzureCloud",
+  EntityFramework: "tech.EntityFramework",
+  Microservices: "tech.Microservices",
+  RestAPI: "tech.RestAPI",
+  ConfigurationDrivenUI: "tech.ConfigurationDrivenUI",
+  DataDrivenUI: "tech.DataDrivenUI",
+  Kubernetes: "tech.Kubernetes",
+  xUnit: "tech.xUnit",
+  AzureDevOps: "tech.AzureDevOps",
+  AngularMaterial: "tech.AngularMaterial",
+  AzureCosmosDB: "tech.AzureCosmosDB",
+  SQLServer: "tech.SQLServer",
+  ReactRouter: "tech.ReactRouter",
+  Scrum: "tech.Scrum",
+  SourceTree: "tech.SourceTree",
+  NHibernate: "tech.NHibernate",
+  MediatR: "tech.MediatR",
+  AspNetMVC: "tech.AspNetMVC",
+  CQRS: "tech.CQRS",
+  MediatorPattern: "tech.MediatorPattern",
+  TSLint: "tech.TSLint",
+  OctopusDeploy: "tech.OctopusDeploy",
+  UIKit: "tech.UIKit",
+  IceLib: "tech.IceLib",
+  IntegrationTesting: "tech.IntegrationTesting",
+  NodeJS: "tech.NodeJS",
+  Express: "tech.Express",
+  GraphQL: "tech.GraphQL",
+  MonoreposNx: "tech.MonoreposNx",
+  CleanArchitecture: "tech.CleanArchitecture",
+  DomainDrivenDesign: "tech.DomainDrivenDesign",
+  EventDrivenDesign: "tech.EventDrivenDesign",
+  SSRSSG: "tech.SSRSSG",
+  GitHubActions: "tech.GitHubActions",
+  CICDPipelines: "tech.CICDPipelines",
+  ReactTestingLibrary: "tech.ReactTestingLibrary",
+  Vitest: "tech.Vitest",
+  PostgreSQL: "tech.PostgreSQL",
+  Vercel: "tech.Vercel",
+  Netlify: "tech.Netlify",
+  Firebase: "tech.Firebase",
+} as const;
+
+export type Tech = (typeof Tech)[keyof typeof Tech];
+
 
 export interface WorkEntry {
   id: string;
@@ -285,7 +311,7 @@ export const workHistory: WorkEntry[] = [
     location: "Amsterdam",
     mode: WorkMode.OnSite,
     language: Language.Dutch,
-    color:"#324e64",
+    color: "#324e64",
     tech: [
       Tech.Angular,
       Tech.TypeScript,
@@ -445,7 +471,7 @@ export const workHistory: WorkEntry[] = [
     location: "Hilversum",
     mode: WorkMode.OnSite,
     language: Language.English,
-    color:"#DB0717",
+    color: "#DB0717",
     tech: [
       Tech.Angular,
       Tech.TypeScript,
